pyproject.toml
[project]
name = "sd-scripts"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.9",
    "pydantic-argparse>=0.10.0",
    "toml==0.10.2",
    "boto3>=1.40.40",
    "tqdm>=4.67.1",
    "requests>=2.32.5",
    "vastai-sdk>=0.1.20",
    "opencv-python>=4.8.1.78",
    "onnx>=1.18.0",
    "onnxruntime-gpu>=1.22.0",
    "huggingface-hub==0.24.5",
    "pandas>=2.3.3",
]

[dependency-groups]
train = [
    "accelerate>=0.30.0",
    "altair==5.5.0",
    "bitsandbytes>=0.47.0",
    "civitdl>=2.1.2",
    "diffusers[torch]==0.25.0",
    "easygui==0.98.3",
    "einops==0.7.0",
    "fairscale>=0.4.13",
    "ftfy==6.1.1",
    "gdown>=5.2.0",
    "imagesize==1.4.1",
    "lion-pytorch==0.0.6",
    "lycoris-lora>=3.2.0.post2",
    "numpy<2",
    "opencv-python==4.8.1.78",
    "prodigyopt==1.0",
    "pytorch-lightning==1.9.0",
    "rich==13.7.0",
    "safetensors==0.4.2",
    "timm>=1.0.19",
    "torchvision>=0.23.0",
    "triton>=3.4.0",
    "voluptuous==0.13.1",
    "xformers>=0.0.32.post2",
    "transformers==4.44.0",
    "torch>=2.8.0",
    "numpy>=1.26.4",
]
